refactor(plugin-eslint): reorganize files
