+++ packages/plugin-eslint/src/lib/meta/transform.ts
@@ -1,15 +1,6 @@
 import type { Audit } from '@code-pushup/models';
-import type { ESLint } from 'eslint';
 import { ruleIdToSlug } from './hash';
-import { RuleData, listRules } from './rules';
-
-export async function listAudits(
-  eslint: ESLint,
-  patterns: string | string[],
-): Promise<Audit[]> {
-  const rules = await listRules(eslint, patterns);
-  return rules.map(ruleToAudit);
-}
+import { RuleData } from './rules';
 
 export function ruleToAudit({ ruleId, meta, options }: RuleData): Audit {
   const name = ruleId.split('/').at(-1) ?? ruleId;
