+++ packages/plugin-eslint/src/lib/runner/index.ts
@@ -1,0 +1,2 @@
+export const RUNNER_OUTPUT_PATH =
+  'node_modules/.code-pushup/eslint/runner-output.json';
